header & list1

+++ ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: owatanab <owatanab@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/21 17:39:32 by owatanab          #+#    #+#             */
-/*   Updated: 2023/06/22 13:52:09 by owatanab         ###   ########.fr       */
+/*   Updated: 2023/07/25 16:49:55 by owatanab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned n)
+void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*a;
 
 	a = s;
